refactor(sidecar): four groups of helpers find the module that owns them
terra/actions.py drops from 3,789 lines to 3,508.

terra/aoi.py takes polygon_from_geojson and parse_kml_coordinates: the study
area, however it arrived. It reads no imagery and decides nothing about the
ground, which is why it sits below every product rather than inside one.

terra/imagery/grid.py takes the three questions asked about the reference grid
once it exists -- its pixel size, its extent, and putting another raster on it.
reproject_mapbiomas_to_grid becomes reproject_to_reference, because MapBiomas
was what its caller passed and not what it does, and the name kept the
photovoltaic siting chain looking like a consumer of the land-cover product.

terra/flood/envelope.py takes the shared-window crop, the AOI reporting mask
and the agreement colouring. terra/energy/siting.py takes compute_siting.

WHAT THE CONTRACT REFUSED. With compute_siting inside the energy slice,
import-linter reported terra.energy.siting -> terra.landcover.mapbiomas: the
function resolved which MapBiomas raster covers the AOI, and fetched it if it
was not on disk. That is the land-cover product's work. compute_siting now
takes a resolved path and the two actions resolve it, which is what an action
is for -- orchestrating across products is the one place that edge is allowed
to exist.

The deferred imports in terra/actions.py are recomputed from the attribute
reads in each function rather than patched: the first attempt inserted them
against line numbers that later edits had moved, and left five functions
importing what they did not use and three using what they had not imported.
ruff reported both halves.

The module alias is ref_grid, not grid: the flood action binds a local named
grid for the window the products were compared on.

Verification: pytest 520 passed, 8 skipped. ruff check clean, three contracts
kept, gofmt, go vet and go test -race clean. A ping through the entry point
answers.
